refactor: segregando metodos para role admin, controllers e DAOs
