+++ src/main/java/br/com/guedes/apartments/controllers/AuthenticationController.java
@@ -3,8 +3,8 @@ package br.com.guedes.apartments.controllers;
 import br.com.guedes.apartments.config.security.TokenGenerateService;
 import br.com.guedes.apartments.models.dto.responses.DefaultResponseDTO;
 import br.com.guedes.apartments.models.dto.responses.LoginTokenResponseDTO;
-import br.com.guedes.apartments.models.dto.requests.RequestLoginDTO;
-import br.com.guedes.apartments.models.dto.requests.RequestRegisterDTO;
+import br.com.guedes.apartments.models.dto.authorization.RequestLoginDTO;
+import br.com.guedes.apartments.models.dto.authorization.RequestRegisterDTO;
 import br.com.guedes.apartments.models.dto.authorization.UserSecurityDetails;
 import br.com.guedes.apartments.models.enums.Message;
 import br.com.guedes.apartments.services.UserService;
