+++ src/main/java/br/com/guedes/apartments/dao/UserDAO.java
@@ -1,7 +1,6 @@
 package br.com.guedes.apartments.dao;
 
 import br.com.guedes.apartments.ConnectionFactory;
-import br.com.guedes.apartments.models.dto.responses.UserFetcherDTO;
 import br.com.guedes.apartments.models.dto.authorization.UserSecurityDetails;
 import br.com.guedes.apartments.models.enums.Role;
 import org.springframework.stereotype.Repository;
@@ -11,9 +10,7 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.Date;
-import java.util.List;
 
 @Repository
 public class UserDAO {
@@ -47,52 +44,7 @@ public class UserDAO {
         }
     }
 
-    public List<UserFetcherDTO> selectAllUsers() {
-        Connection conn = connectionFactory.getConnection();
-        List<UserFetcherDTO> usersList = new ArrayList<>();
-        String sql = "SELECT id, cpf, password, name, role, creation FROM users";
-
-        PreparedStatement preparedStatement = null;
-        try {
-            preparedStatement = conn.prepareStatement(sql);
-            ResultSet resultSet = preparedStatement.executeQuery();
-            while (resultSet.next()) {
-                UserFetcherDTO user = mapUser(resultSet);
-                usersList.add(user);
-            }
-            preparedStatement.close();
-            conn.close();
-        } catch (SQLException e) {
-            throw new RuntimeException("Error when searching for users", e);
-        }
-
-        return usersList;
-    }
-
-    public UserFetcherDTO selectUserForName(String username) {
-        Connection conn = connectionFactory.getConnection();
-        String sql = "SELECT id, cpf, name, role, creation FROM users WHERE name = ?";
-        UserFetcherDTO userResponse = null;
-
-        PreparedStatement preparedStatement = null;
-        try {
-            preparedStatement = conn.prepareStatement(sql);
-            preparedStatement.setString(1, username);
-            ResultSet resultSet = preparedStatement.executeQuery();
-            while (resultSet.next()) {
-                userResponse = mapUser(resultSet);
-            }
-
-            preparedStatement.close();
-            conn.close();
-        } catch (SQLException e) {
-            throw new RuntimeException("Error when searching for name user", e);
-        }
-
-        return userResponse;
-    }
-
-    public UserSecurityDetails selectUserForCPF(String cpf) {
+    public UserSecurityDetails selectUserDetailsForCPF(String cpf) {
         Connection conn = connectionFactory.getConnection();
         String sql = "SELECT cpf, password, role FROM users WHERE cpf = ?";
         UserSecurityDetails userResponse = null;
@@ -113,15 +65,6 @@ public class UserDAO {
         }
 
         return userResponse;
-    }
-
-    private UserFetcherDTO mapUser(ResultSet rs) throws SQLException {
-        return new UserFetcherDTO(
-                rs.getLong("id"),
-                rs.getString("cpf"),
-                rs.getString("name"),
-                rs.getString("role"),
-                rs.getString("creation"));
     }
 
     private UserSecurityDetails mapUserForUserDetails(ResultSet rs) throws SQLException {
