+++ src/main/java/br/com/guedes/apartments/models/dto/authorization/RequestLoginDTO.java
@@ -1,3 +1,3 @@
-package br.com.guedes.apartments.models.dto.requests;
+package br.com.guedes.apartments.models.dto.authorization;
 
 public record RequestLoginDTO(String cpf, String password) {}
